+++ src/Config/Enum/CompressionType.php
@@ -6,6 +6,7 @@ namespace Anktx\Kafka\Client\Config\Enum;
 
 enum CompressionType: string
 {
+    case none = 'none';
     case snappy = 'snappy';
     case gzip = 'gzip';
     case lz4 = 'lz4';